virtio: separate virtio-fs into separate file
