+++ src/arch/x86_64/kernel/mod.rs
@@ -32,6 +32,7 @@ pub mod systemtime;
 #[cfg(feature = "vga")]
 mod vga;
 pub mod virtio;
+pub mod virtio_fs;
 
 use arch::x86_64::kernel::percore::*;
 use arch::x86_64::kernel::serial::SerialPort;
